(FACT-3468) Initialize aggregate log
Previously if an aggregate fact was evaluated twice it would attempt to log a
warning but its `log` method was never defined, leading to NoMethodError.

It appears this was broken in ee582c5714228ceff61a1105cf439cb4ca53ff12 when the
Legacy logger was removed. This commit adds a private `log` method to lazily
create the log as is done elsewhere and adds a test.

Since string literals are frozen, we need to dup prior to modifying it.