(FACT-3468) Get Facter global logger instead of stubbing it
Previously, we were overwriting the global logger with an instance spy. In some
cases we were not clearing it when we were done or we set it to a new instance
of Logger.new(STDOUT). In the first case, the instance spy could leak across to
other tests. In the second case, we've replaced the global logger with a
different instance instead of replacing it with the original instance.

All of this can be avoided by just getting the logger and setting expectations
on it. Since it's a "class instance variable" we have to use `:@@logger` syntax.

Since the global logger is not a spy, then we have to set the expectation before
calling the method that triggers the message to be sent such as:

        expect(logger).to receive(:debug).with('...')
        log.debug('debug_message')