(FACT-3468) Pass Log to execution methods
Previously, we were stubbing Facter::Log.new to return our instance_spy. But we
can just pass the instance_spy as an keyword argument to the `execute` method.

Confusingly the `execute` method takes a `logger` keyword argument, but
internally it creates a `Facter::Log` object, so the argument should really be
called `log`, but that's part of the execution API and can't be changed.