Set test to pending due to JRuby bug
JRuby 9.4.7.0, which the latest version of Puppet Server 8.x uses,
does not properly parse IP addresses that mix IPv6 and IPv4 syntax
(e.g. ::ffff:192.0.2.128). Because of this, we have filed a bug usptream
and have set a test that uses that style of address to pending.

See https://github.com/jruby/jruby/issues/8248