(packaging) Add rake tasks for building nightly gem
The nightly gem's version (aka extended_dot_version) is generated from git
describe as the last tag, plus some number of commits and abbreviated git ref:

    4.7.0-36-g2ab655b

Additionally, if there are changes in the worktree, then we add 'dirty'
to the version.

    4.7.0-36-g2ab655b-dirty

The "extended_dot_version" is added to a temporary gemspec that's used
to build the gem. That version determines the resulting filename, e.g.

    facter-4.7.0-36-g2ab655b.gem

Neither the packaging repo nor this commit update the version in `version.rb`,
so running facter --version may return the "version to be released in the
future".