Use bundler-cache to speed up CI
See puppetlabs/puppet@86820eee25e428f7e667bf9529b97c32e6575d9b

This intentionally doesn't use bundler-cache for the mend workflow, since we
want the latest versions of all of our dependencies that still satisfy our
constraints.

This doesn't update the acceptance_tests workflow because it does a bundle
install for both the top-level Gemfile and the one in the acceptance dir.