Add error handling for older lsblk
The partitions resolver uses `lsblk --version` to determine the version
of lsblk and what features it has available to it. The --version flag
was not added until lsblk 2.22 in 2012, so versions prior to that will
respond with `lsblk: unrecognized option '--version'`.

Puppet still supports older operating systems like SUSE Enterprise Linux
11 that may ship versions of lsblk that do not support --version.

This commit updates the partitions resolver to handle errors when
determining the version of lsblk and to fall back to blkid when lsblk is
too old to support --version.