Install gems without documentation group
As part of our presuite actions, we verify that we can run Facter as a
standalone Ruby application. During this, we run a `bundle install`,
which installs documentation dependencies.

It is a known issue that the manpage generation library Facter uses,
ronn, is out-of-date and errors on modern compilers. See:
https://github.com/puppetlabs/facter/issues/2703

This commit updates our presuit action to omit installing documentation
dependencies. This also removes the option of omitting development
dependencies, as there is no such group in Facter's gemfile.