(FACT-3488) Ignore unknown container types
Prior to commit 7bc38ccb58e6abf66d6141a3b2e9f60eb0077d71, facter checked for
docker and lxc in `/proc/1/cgroups` and fell back to the `container` environment
variable in `/proc/1/environ`.

The above commit reversed the order and labeled unknown container types as
'container_other'. This broke the VirtualDetecter, because it calls the
`containers` resolver and assumes it can fallback to other methods like
check_freebsd, etc

https://github.com/puppetlabs/facter/blob/0656d9a34ce4790129f1fd6eba5bb4d49a9b9ad1/lib/facter/util/facts/posix/virtual_detector.rb#L11

The regression was noticed and corrected for Illumox LX in commit
31d32da513aae094bdbcc011155c2e247be11284, but we didn't realize other
forms of virtualization were affected.

This commit restores the original behavior of checking `/proc/1/cgroups` first,
then `/proc/1/environ`. If a container type is not recognized, we log a debug
message and continue, so that VirtualDetector works as expected.