(FACT-3495) Add missing chassis types
Currently, chassis types exist in a hash called types in the function
chassis_to_name in dmi.rb. Each type's position and ordering in the hash
corresponds with its corresponding number/hex. This commit adds the following
missing chassis types to the hash:
  - IoT Gateway (byte value: 21h)
  - Embedded PC (byte value: 22h)
  - Mini PC (byte value: 23h)
  - Stick PC (byte value: 24h)

For more information on chassis types, see:
https://www.dmtf.org/sites/default/files/standards/documents/DSP0134_3.8.0.pdf