Don't load spec/**/*.rb as custom facts (again)
Commit c3350f82f8bb6b74fb6f18bfbfa147bed759b32b removed the `facter/spec`
directory from the load path so that Facter wouldn't attempt to load custom
facts in "spec/**/*.rb". However, when we forked the repo, the name changed
to `facter-private`, reintroducing the problem.

This commit resolves the absolute path to the spec directory, so we're not
sensitive to the name of the git checkout directory.