(FACT-3497) Facter should not fail when binary data is detected
This commit updates LegacyFacter::Util::Normalization.normalize_string to
accept fact values with ASCII 8-BIT without immediately failing. When a fact
value with ASCII 8-BIT is detected but can be force encoded to UTF-8, Facter
will log a debug message with the fact name and value. Facter will only error
(with the fact name & value) when the fact value cannot be successfully force
encoded to UTF-8.

This commit also updates all the normalize methods in
LegacyFacter::Util::Normalization to take in 2 parameters now: the fact name &
value so the debug and error messages can include the fact name. This helps
users more easily figure out which fact is causing issues and has binary.